Split out + renamespace edge effect related classes
