+++ osu.Framework/Graphics/Containers/CompositeDrawable_DrawNode.cs
@@ -10,68 +10,11 @@ using osuTK;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Graphics.Colour;
 using System;
+using osu.Framework.Graphics.Effects;
 using osu.Framework.Graphics.OpenGL.Vertices;
 
 namespace osu.Framework.Graphics.Containers
 {
-    /// <summary>
-    /// Types of edge effects that can be applied to <see cref="CompositeDrawable"/>s.
-    /// </summary>
-    public enum EdgeEffectType
-    {
-        None,
-        Glow,
-        Shadow,
-    }
-
-    /// <summary>
-    /// Parametrizes the appearance of an edge effect.
-    /// </summary>
-    public struct EdgeEffectParameters : IEquatable<EdgeEffectParameters>
-    {
-        /// <summary>
-        /// Colour of the edge effect.
-        /// </summary>
-        public SRGBColour Colour;
-
-        /// <summary>
-        /// Positional offset applied to the edge effect.
-        /// Useful for off-center shadows.
-        /// </summary>
-        public Vector2 Offset;
-
-        /// <summary>
-        /// The type of the edge effect.
-        /// </summary>
-        public EdgeEffectType Type;
-
-        /// <summary>
-        /// How round the edge effect should appear. Adds to the <see cref="CompositeDrawable.CornerRadius"/>
-        /// of the corresponding <see cref="CompositeDrawable"/>. Not to confuse with the <see cref="Radius"/>.
-        /// </summary>
-        public float Roundness;
-
-        /// <summary>
-        /// How "thick" the edge effect is around the <see cref="CompositeDrawable"/>. In other words: At what distance
-        /// from the <see cref="CompositeDrawable"/>'s border the edge effect becomes fully invisible.
-        /// </summary>
-        public float Radius;
-
-        /// <summary>
-        /// Whether the inside of the EdgeEffect rectangle should be empty.
-        /// </summary>
-        public bool Hollow;
-
-        public bool Equals(EdgeEffectParameters other) =>
-            Colour.Equals(other.Colour) &&
-            Offset == other.Offset &&
-            Type == other.Type &&
-            Roundness == other.Roundness &&
-            Radius == other.Radius;
-
-        public override string ToString() => Type != EdgeEffectType.None ? $@"{Radius} {Type}EdgeEffect" : @"EdgeEffect (Disabled)";
-    }
-
     public partial class CompositeDrawable
     {
         /// <summary>
