+++ osu.Framework/Graphics/Effects/EdgeEffectType.cs
@@ -1,0 +1,17 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Graphics.Containers;
+
+namespace osu.Framework.Graphics.Effects
+{
+    /// <summary>
+    /// Types of edge effects that can be applied to <see cref="CompositeDrawable"/>s.
+    /// </summary>
+    public enum EdgeEffectType
+    {
+        None,
+        Glow,
+        Shadow,
+    }
+}
